requirements.txt
click~=7.1
pydantic~=1.6
PyGithub~=1.51
requests~=2.24
pip-download @ https://github.com/saklar13/pip-download/archive/master.zip
